requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 25.5.1
accelerate=1.11.0=pypi_0
aiofiles=25.1.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.13.2=pypi_0
aiosignal=1.4.0=pypi_0
altair=5.5.0=pypi_0
annotated-doc=0.0.3=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.11.0=pypi_0
appnope=0.1.4=pyhd8ed1ab_1
asttokens=3.0.0=pyhd8ed1ab_1
async-timeout=4.0.3=pypi_0
attrs=25.4.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=5.0.0=pypi_0
beautifulsoup4=4.14.2=pypi_0
blinker=1.9.0=pypi_0
build=1.3.0=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2025.11.4=hca03da5_0
cachetools=6.2.1=pypi_0
certifi=2025.10.5=pypi_0
cffi=2.0.0=pypi_0
charset-normalizer=3.4.4=pypi_0
chroma-hnswlib=0.7.3=pypi_0
chromadb=0.4.15=pypi_0
click=8.3.0=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.2.3=pyhe01879c_0
contourpy=1.3.2=pypi_0
cryptography=46.0.3=pypi_0
cycler=0.12.1=pypi_0
dataclasses-json=0.6.7=pypi_0
debugpy=1.8.16=py310h50f4ffc_1
decorator=5.2.1=pyhd8ed1ab_0
deprecated=1.3.1=pypi_0
distro=1.9.0=pypi_0
durationpy=0.10=pypi_0
effdet=0.4.1=pypi_0
emoji=2.15.0=pypi_0
exceptiongroup=1.3.0=pyhd8ed1ab_0
executing=2.2.1=pyhd8ed1ab_0
expat=2.7.3=h982b769_0
fastapi=0.121.0=pypi_0
filelock=3.20.0=pypi_0
filetype=1.2.0=pypi_0
flatbuffers=25.9.23=pypi_0
fonttools=4.60.1=pypi_0
frozenlist=1.8.0=pypi_0
fsspec=2025.10.0=pypi_0
gitdb=4.0.12=pypi_0
gitpython=3.1.45=pypi_0
google-api-core=2.28.1=pypi_0
google-auth=2.43.0=pypi_0
google-cloud-vision=3.11.0=pypi_0
googleapis-common-protos=1.72.0=pypi_0
grpcio=1.76.0=pypi_0
grpcio-status=1.76.0=pypi_0
h11=0.16.0=pypi_0
hf-xet=1.2.0=pypi_0
html5lib=1.1=pypi_0
httpcore=1.0.9=pypi_0
httptools=0.7.1=pypi_0
httpx=0.28.1=pypi_0
httpx-sse=0.4.3=pypi_0
huggingface-hub=0.36.0=pypi_0
humanfriendly=10.0=pypi_0
idna=3.11=pypi_0
importlib-metadata=8.7.0=pyhe01879c_1
importlib-resources=6.5.2=pypi_0
ipykernel=7.1.0=pyh5552912_0
ipython=8.37.0=pyh8f84b5b_0
jedi=0.19.2=pyhd8ed1ab_1
jinja2=3.1.6=pypi_0
jiter=0.11.1=pypi_0
joblib=1.5.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.25.1=pypi_0
jsonschema-specifications=2025.9.1=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.9.1=pyhc90fa1f_0
kiwisolver=1.4.9=pypi_0
krb5=1.21.3=h237132a_0
kubernetes=34.1.0=pypi_0
langchain=1.0.5=pypi_0
langchain-classic=1.0.0=pypi_0
langchain-community=0.4.1=pypi_0
langchain-core=1.0.4=pypi_0
langchain-openai=1.0.2=pypi_0
langchain-text-splitters=1.0.0=pypi_0
langdetect=1.0.9=pypi_0
langgraph=1.0.2=pypi_0
langgraph-checkpoint=3.0.1=pypi_0
langgraph-prebuilt=1.0.2=pypi_0
langgraph-sdk=0.2.9=pypi_0
langsmith=0.4.41=pypi_0
libcxx=20.1.8=hd7fd590_1
libedit=3.1.20250104=pl5321hafb1f1b_0
libffi=3.4.4=hca03da5_1
libsodium=1.0.20=h99b78c6_0
libzlib=1.3.1=h5f15de7_0
lxml=6.0.2=pypi_0
markdown-it-py=4.0.0=pypi_0
markupsafe=3.0.3=pypi_0
marshmallow=3.26.1=pypi_0
matplotlib=3.10.7=pypi_0
matplotlib-inline=0.2.1=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
ml-dtypes=0.5.3=pypi_0
mmh3=5.2.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.7.0=pypi_0
mypy-extensions=1.1.0=pypi_0
narwhals=2.10.2=pypi_0
ncurses=6.5=hee39554_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pypi_0
nltk=3.9.2=pypi_0
numpy=2.2.6=pypi_0
oauthlib=3.3.1=pypi_0
olefile=0.47=pypi_0
omegaconf=2.3.0=pypi_0
onnx=1.19.1=pypi_0
onnxruntime=1.23.2=pypi_0
openai=2.7.1=pypi_0
opencv-python=4.12.0.88=pypi_0
openssl=3.5.4=h5503f6c_0
opentelemetry-api=1.38.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.38.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.38.0=pypi_0
opentelemetry-proto=1.38.0=pypi_0
opentelemetry-sdk=1.38.0=pypi_0
opentelemetry-semantic-conventions=0.59b0=pypi_0
orjson=3.11.4=pypi_0
ormsgpack=1.12.0=pypi_0
overrides=7.7.0=pypi_0
packaging=25.0=pyh29332c3_1
pandas=2.3.3=pypi_0
parso=0.8.5=pyhcf101f3_0
pdf2image=1.17.0=pypi_0
pdfminer-six=20251107=pypi_0
pexpect=4.9.0=pyhd8ed1ab_1
pi-heif=1.1.1=pypi_0
pickleshare=0.7.5=pyhd8ed1ab_1004
pikepdf=10.0.0=pypi_0
pillow=12.0.0=pypi_0
pinecone-client=3.0.0=pypi_0
pip=25.2=pyhc872135_1
platformdirs=4.5.0=pyhcf101f3_0
posthog=5.4.0=pypi_0
prompt-toolkit=3.0.52=pyha770c72_0
propcache=0.4.1=pypi_0
proto-plus=1.26.1=pypi_0
protobuf=6.33.0=pypi_0
psutil=7.1.3=pypi_0
ptyprocess=0.7.0=pyhd8ed1ab_1
pulsar-client=3.8.0=pypi_0
pure_eval=0.2.3=pyhd8ed1ab_1
pyarrow=21.0.0=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
pybase64=1.4.2=pypi_0
pycocotools=2.0.10=pypi_0
pycparser=2.23=pypi_0
pydantic=2.12.4=pypi_0
pydantic-core=2.41.5=pypi_0
pydantic-settings=2.11.0=pypi_0
pydeck=0.9.1=pypi_0
pygments=2.19.2=pyhd8ed1ab_0
pyparsing=3.2.5=pypi_0
pypdf=6.1.3=pypi_0
pypdfium2=5.0.0=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.2.0=pypi_0
python=3.10.19=hf701271_0
python-dateutil=2.9.0.post0=pyhe01879c_2
python-docx=1.2.0=pypi_0
python-dotenv=1.2.1=pypi_0
python-iso639=2025.2.18=pypi_0
python-magic=0.4.27=pypi_0
python-multipart=0.0.20=pypi_0
python-oxmsg=0.0.2=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.3=pypi_0
pyzmq=27.1.0=py310h5a8c52c_1
rapidfuzz=3.14.3=pypi_0
readline=8.3=h0b18652_0
referencing=0.37.0=pypi_0
regex=2025.11.3=pypi_0
requests=2.32.5=pypi_0
requests-oauthlib=2.0.0=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=14.2.0=pypi_0
rpds-py=0.28.0=pypi_0
rsa=4.9.1=pypi_0
safetensors=0.6.2=pypi_0
scikit-learn=1.7.2=pypi_0
scipy=1.15.3=pypi_0
sentence-transformers=5.1.2=pypi_0
setuptools=80.9.0=py310hca03da5_0
shellingham=1.5.4=pypi_0
six=1.17.0=pyhe01879c_1
smmap=5.0.2=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.8=pypi_0
sqlalchemy=2.0.44=pypi_0
sqlite=3.51.0=hab6afd1_0
stack_data=0.6.3=pyhd8ed1ab_1
starlette=0.49.3=pypi_0
streamlit=1.51.0=pypi_0
sympy=1.14.0=pypi_0
tenacity=9.1.2=pypi_0
threadpoolctl=3.6.0=pypi_0
tiktoken=0.12.0=pypi_0
timm=1.0.22=pypi_0
tk=8.6.15=hcd8a7d5_0
tokenizers=0.22.1=pypi_0
toml=0.10.2=pypi_0
tomli=2.3.0=pypi_0
torch=2.9.0=pypi_0
torchvision=0.24.0=pypi_0
tornado=6.5.2=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.57.1=pypi_0
typer=0.20.0=pypi_0
typing-inspect=0.9.0=pypi_0
typing-inspection=0.4.2=pypi_0
typing_extensions=4.15.0=pyhcf101f3_0
tzdata=2025.2=pypi_0
unstructured=0.18.18=pypi_0
unstructured-client=0.42.3=pypi_0
unstructured-inference=1.1.1=pypi_0
unstructured-pytesseract=0.3.15=pypi_0
urllib3=2.3.0=pypi_0
uvicorn=0.38.0=pypi_0
uvloop=0.22.1=pypi_0
watchfiles=1.1.1=pypi_0
wcwidth=0.2.14=pyhd8ed1ab_0
webencodings=0.5.1=pypi_0
websocket-client=1.9.0=pypi_0
websockets=15.0.1=pypi_0
wheel=0.45.1=py310hca03da5_0
wrapt=2.0.1=pypi_0
xxhash=3.6.0=pypi_0
xz=5.6.4=h80987f9_1
yarl=1.22.0=pypi_0
zeromq=4.3.5=h888dc83_9
zipp=3.23.0=pyhd8ed1ab_0
zlib=1.3.1=h5f15de7_0
zstandard=0.25.0=pypi_0
